// src/types.ts
export type Language = {
    name: LangNames;
    compiler: string;
    args: string[];
    skipCompile: boolean;
};

export type LangNames =
    | 'python'
    | 'ruby'
    | 'c'
    | 'cpp'
    | 'cc'
    | 'cxx'
    | 'rust'
    | 'java'
    | 'js'
    | 'go'
    | 'hs'
    | 'csharp'
    | 'cangjie';
// ...
export type Run = {
    stdout: string;
    stderr: string;
    code: number | null;
    signal: string | null;
    time: number;
    timeOut: boolean;
};